Initial acceptance tests with Wallaby (#12)
* Set up 

Except phantoms doesn’t support promises, yo

* Add Promise polypill for phantomjs

https://github.com/github/fetch#installation

* Fixup install

* Initial acceptance tests

* Update cadetship’s version of phantoms used

Their default is 1.9.7, but we need at least 2.1.1

* Specify node version and npm install

* Update wallaby

* Whoops, fix typo

* Let’s try hound?

* Revert "Let’s try hound?"

This reverts commit faf9038cfec99c42c037874028317e3498460d7c.

* Install phantoms after node?

* Negative test

* Fail earlier with assert_has

* npm install --save

Maybe? If so, thanks https://elixirforum.com/t/phoenix-and-brunch/1786/3

* Revert "npm install --save"

This reverts commit 2efc58b9f4b5158eb107a3933245505ce47eb072.

* Have Wallaby fail on JS errors

The README makes it sound like this is the default (they say to set it to `false` if you *don’t* want it to fail on JS errors) but I’m seeing I need to set it to `true` to see JS errors.

* Specs passing locally—remove smoke test

* We need to build assets manually in test

* New pg_dump version

* Add promise-polyfill to package.json

Because brunch says it’s necessary

* Fix prop-types deprecation

* Let’s forget about wallaby JS errors for now

A React PropTypes deprecation warning for a library I’m using is setting this off. I’ll update that separately.