Move an mis-scoped expectation
This spec is failing locally because at the point in which we call `refute_has`, we’ve scoped the `css` selector to within `.ActionList-item` thanks to `find` on line 31. By re-scoping `refute_has` to the original `session`, we should be looking for (and expecting not to find) `.ActionList-item` on the entire page, not within the removed `.ActionList-item` element.