Move CheckIn finding and creation into Habits module
I think schema modules should be responsible for defining the data model, but not necessarily querying records from the DB. Since the context module already does that for habits, it seems reasonable that it does so for check-ins too.

Fewer imports and aliases overall seems like a good thing, though I suppose this could be taken too far.